Fixed missing border for tab widget in Projects mode
Usually for tab widgets we disabled the border of the first tab, since
the tabs are generally aligned to the single pixel black splitters that
we used in Creator (like in the case of dock widgets or application
output).

Since this particular tab widget is not aligned to such a splitter, it
should have its own border which can be achieved by disabling the
custom style on it.

Change-Id: I8c7af60847afb5cd73ecd8c39a834ed63298ec76
Reviewed-on: http://codereview.qt.nokia.com/122
Reviewed-by: Jens Bache-Wiig