QmlJSInspector: Fixed multi-selection bug
Nothing useful can be done with multi-selection at the moment, but this
bug was introduced in 8d0601515ebf54568609b2c7e73a8a91c63fdc26 and is
better to fix straight away.

What happened was that when multiple items are selected in the observed
QML application, Qt Creator would pick the first and send that back as
the new selection. This made it impossible to select more than one item.

Task-number: QTCREATORBUG-3426
Change-Id: I68283480af5e0a21e32e51a49848ff9a809ed5ec
Reviewed-on: http://codereview.qt.nokia.com/151
Reviewed-by: Kai Koehne <kai.koehne@nokia.com>