QmlDesigner.model: refactoring for nodeSource
I changed customParserSource into nodeSource and use it also for
components. This means we do not use auxiliaryData anymore with some magic
__component_data identifier.
To distinguish between CustumParserSource and ComponentSource I introduced
the enum NodeSourceType.
Also in this patch I added auxiliaryData in the creation of a node. This was useful
for prototyping and will help with prototyping/testing in the future.

Change-Id: I2152c26c0c767f869f7dce7209abf43f594fd2ad
Reviewed-on: http://codereview.qt.nokia.com/399
Reviewed-by: Marco Bubke <marco.bubke@nokia.com>