Add a toolchain combo box to the cmake run page
Selecting the toolchain is a configure time setting for cmake. So
that's the right place, except for us the debugger also depends
on the toolchain. Switching the debugger thus involves reconfiguring.

Also due to the cmake cache in most cases reconfiguring with a different
toolchain selected won't do anything.

Cmake and creator's concept of toolchains clash.

Task-Nr: QTCREATORBUG-4898

Change-Id: Ibf84029d7008046891750a91666d0e3e4711ec69
Reviewed-on: http://codereview.qt.nokia.com/477
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Daniel Teske <daniel.teske@nokia.com>