Revert "Completion: Apply perfect match whenever not invoked explicitly"
This reverts commit 4959f53d2c54515de73cb40c8363cd7ce6d24e03.

It seems that usability is better when upon a perfect match the completion
popup stays when triggered either explicitly or by an activation character.
After all, in general one should need to type something else (semicolon,
parenthesis, etc.) anyway for the case of an activation character.

This is also consistent with the behavior from previous releases.

Change-Id: I1d50230affb1d0642f8f8bfd5cbac69dc52e44bd
Reviewed-on: http://codereview.qt.nokia.com/536
Reviewed-by: Thorbjørn Lindeijer <thorbjorn@lindeijer.nl>