CodeAssist: Relax a bit premature match
In a relatively recent commit (99fee33fd1a5e82b7be3d17a74229efaaf7571ef)
premature match was setup to work only on explicit invocations. However
this has shown to be too strict. Therefore, it's now a bit relaxed (not
as much as before): If the current item ends with the typed character we
assume  this is intenional - general premature matches do not end with
identifier characters anyway.

Note: Leaving only the last character as the condition is not an option
since sometimes they are not shown in the proposal description.

This patch also lifts the text method to IAssistProposalItem since
it's reasonable the every proposal must have at least some description.

Change-Id: I2ef7de2b7f978ce7059cce50175137a03315f495
Reviewed-on: http://codereview.qt.nokia.com/675
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Christian Kamm <christian.d.kamm@nokia.com>