VCS: Improve tab width handling in VCS editor
Use settings for generic files that do not belong to any project for
e.g. git diff. That way a user can actually override the tab width
which used to be hardcoded to 8.

This is not perfect: Actually we should use the tab width defined for
the file type in the project it belongs to. That does require
significantly more work though.

Task-number: QTCREATORBUG-5178
Change-Id: I509b030e09c33a74ecd529294bf08afec21b2121
Reviewed-on: http://codereview.qt.nokia.com/768
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Tobias Hunger <tobias.hunger@nokia.com>