QmlDesigner: Fix painted bounding rectangle for step childs
We used only the bounding rectangle of the item under our control. But child
items which are created in a component can be out side of the bounding
rectangle. To fix compute a combined bounding rectangle of the item and all
stepchildren. Stepchildren described the child not create directly by us.

Change-Id: I7f25f112aa5eb2f80cfccec5d358221ab5e35d9b
Reviewed-on: http://codereview.qt.nokia.com/839
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@nokia.com>