qmldump: Fix dumping with -path on Mac.
If the current working directory was a direct parent of the qmldir path
the exported modules had the path as the module URI on macs.

Also changes the QtQuick export back to 1.0 to make it work with
Qt 4.7.3. The version of that import statement does not actually matter
as long as it's valid.

Change-Id: Ib702356062bd239b1ad807c2592f8a69d6fde38f
Reviewed-on: http://codereview.qt.nokia.com/896
Reviewed-by: Kai Koehne <kai.koehne@nokia.com>
(cherry picked from commit dc3e803614e145baa425b22162fc39b286d16a52)