Don't jump to the wrong qml file in insource build.
Doing an insource build on Mac will copy the qml file(s) into the
applications resources folder, which still is located under the project
root. In that case it was assumed to be the original file. The patch
assumes that for files in the project root with
".app/Contents/Resources" in the path will need the magic to happen.

Change-Id: I25ffea8a1be7caff5313d03590b4094cb3429492
Reviewed-on: http://codereview.qt.nokia.com/1698
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Christiaan Janssen <christiaan.janssen@nokia.com>
Reviewed-by: Kai Koehne <kai.koehne@nokia.com>