QmlDesigner: fix in model for meta info
We have to use metaInfoProxyModel() to allow forwarding.

Since property.parentModelNode().model() might be null in
propertyIsComponentType().

I also provide a fix for this. Before calling ->metaInfo on a 0 model was
safe, since metaInfo is non virtual and the constructor and isSubCLassOf()
were safe and had checks.
This was still wrong of course.

Change-Id: If4a2371e4dceeccba7f3ddde7236a7240ed7b92a
Reviewed-on: http://codereview.qt.nokia.com/1707
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@nokia.com>