QmlJsDbg: Fix compilation for Symbian
Despite of '-k' passed to make, the 'all' target might still fail prematurely
because make doesn't forward the '-k' to sbs. Anyway, now that we
know the exact toolchain we can do better and only compile the
targets that fit to the toolchain.

If not Symbian, use the 'all' target everywhere, which ensures
we're both compiling debug and release versions of all helpers.

Change-Id: I0c3b699724472c8c2035d2c20bc451b47a301d1d
Reviewed-on: http://codereview.qt.nokia.com/2159
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Alessandro Portale <alessandro.portale@nokia.com>