JSDebugger: Only hit breakpoints in user code
Check that the topmost stack entry is a user defined ScriptFunction.
This avoids hitting the anonymous functions used for bindings, e.g.

onClicked: Qt.quit()

leads to script code

(function onClicked() { Qt.quit(); })

which will be hit twice for the debugger: Once for the function call
itself, then for the execution of Qt.quit().

Change-Id: I4cb374782c93a26d97d4a717ce67d1fb2f6df438
Task-number: QTCREATORBUG-5090
Reviewed-on: http://codereview.qt.nokia.com/2746
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Christiaan Janssen <christiaan.janssen@nokia.com>