re-add LD_LIBRARY_PATH script
we need to override RUNPATHs from possible plugins.

due to d7d23226, this should not re-introduce QTCREATORBUG-1646.

unlike in the previous incarnation, we don't replace the binary with the
script, but give the script an extension. this is nicer for packagers
and less confusing in creator's own run configurations. the .desktop
files created by our installers need to be adjusted for that.

Task-number: QTCREATORBUG-5565
Change-Id: Icd2fa55456754a05257376b8288e8bdf423c62db
Reviewed-on: http://codereview.qt.nokia.com/3180
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Daniel Molkentin <daniel.molkentin@nokia.com>
Reviewed-by: Christian Kamm <christian.d.kamm@nokia.com>