DebuggingHelper: Update UI even if compilation failed
This fixes all the buttons staying in disabled mode after an (unsuccessful)
compilation. Problem was that in the end, the finished() call informed
the option page that the build was over, while the updateQtVersions()
call updated the UI. The patch fixes the order, but also updates the UI
as reaction to the finished call, because relying on the delivery order of
two independent signals via Qt::QueuedConnection is fragile per se.

Change-Id: Ifdd03e3cb9d9bf96b30053e5e8aa0a3f0232ab1c
Task-number: QTCREATORBUG-5906
Reviewed-on: http://codereview.qt.nokia.com/3485
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Christiaan Janssen <christiaan.janssen@nokia.com>
Reviewed-by: Daniel Teske <daniel.teske@nokia.com>