Code assist: Fix doxygen completion
When the assist interface is detached the correct thing to do
is to clone the document, not simply to copy the text content.
Then data like the user state is preserved.

For the particular issue below, the problem is that the simple
lexer doesn't have the correct initial state and thus cannot
identify its inside a comment.

Task-number: QTCREATORBUG-6061
Change-Id: I59380eeb229e322bafb1e799d649df90327c0f54
Reviewed-on: http://codereview.qt-project.org/4912
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Christian Kamm <christian.d.kamm@nokia.com>