Find: Change a few accelerators
Since the find widget is no longer a dialog, a few accelerators
"conflict" with others from the main menu and global Creator's.
This doesn't really stop them from working, but it would be better
to have different ones. The new values are not the most obvious
ones, but there's not too many options.

Task-number: QTCREATORBUG-6114
Change-Id: Idbc885abad9b78dad7685568a9524066dc075694
Reviewed-on: http://codereview.qt-project.org/5147
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Eike Ziller <eike.ziller@nokia.com>