RemoteLinux: Add "custom command" deployment step.
It might be useful for developers to run a command before or after
uploading or even have a remote command *be* the "deployment" step.
Also: Stop trying to display the "deployment possible" state in the
widget. It's generally hard to do that correctly and it also interferes
with thread safety.

Change-Id: I28aaf368a645ee14156a165c606dbc17c3063e82
Reviewed-on: http://codereview.qt-project.org/5469
Reviewed-by: Christian Kandeler <christian.kandeler@nokia.com>