WelcomePage: create 2 instances of GettingStartedWelcomePage
GettingStartedWelcomePage gets instanciated twice now.

The new instance does show the examples and has the title:
"Demos and Examples".

The ExamplesListModel is shared between the two instances and is
a static member.
The method examplesModel() is used to access the ExamplesListModel and
lazily creates one.

QtSupportPlugin creates the two different instances.

Change-Id: Ie5bf9ce8b25aeeacb88731ce1d3179004175fd2f
Reviewed-on: http://codereview.qt-project.org/5706
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: hjk <qthjk@ovi.com>