ProjectExplorer: Change how .shared settings work
After discussing the subject we have opted to not automatically track
and synchronize shared settings. The current use case is that one
needs to manually created/edit a .shared file and place it under the
project's directory. We then identify such settings and merge then
into the user ones. Whenever the user changes such a setting in a way
it differs from the .shared value we mark it as a sticky setting.

Change-Id: I1a7196acea5fb76c3ba21ce950a7bfd1aa774da3
Reviewed-on: http://codereview.qt-project.org/5985
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Eike Ziller <eike.ziller@nokia.com>