Add support to force skip deploy before running or debugging
When we enabled the option "Always deploy project before running it",
and try to running or debugging a big project, e.g. QtCreator.
We need waitting a long time for deploy it.
But in fact, we not need this in every time.
So we can add two menu items:
"Run Without Deployment" and "Debug Without Deployment".
We can running or debugging directly.

Change-Id: I6ba92aa0cfa792cef923012856b79c29a672252c
Reviewed-by: Daniel Teske <daniel.teske@nokia.com>
Reviewed-by: Tobias Hunger <tobias.hunger@nokia.com>