turn doc installation into an explicit target
otherwise a "make install" without a prior "make docs" will fail.

we cannot make install depend on docs, because the docs target does not
(and cannot reasonably) have proper dependency tracking, so it would be
rebuilt on every install attempt.
we could make install depend on the presence of the docs files without
checking their timestamps, but that would be dirty.

Task-number: QTCREATORBUG-6245
Change-Id: I35a9b8110f8f12760a338b5b100d69e226d3c2e8
Reviewed-by: Daniel Molkentin <daniel.molkentin@nokia.com>