Bazaar,Mercurial: fix update of the commit button
There was a bug with the initial display of the submit editor.
The submit action was not disabled (description field is empty) and
thus the user was allowed to commit changes without entering a
description. Empty commit message causes error for most VCS (if not all).

Technically, Utils::SubmitEditorWidget::registerActions() must be called
before init of the widget (eg call to BazaarPlugin::setFields()), because
SubmitEditorWidget::updateActions() won't have any chance to be called.

Change-Id: I154d6c807e0943b98abcb6222d8f57ec5421181c
Reviewed-by: Tobias Hunger <tobias.hunger@nokia.com>
Reviewed-by: Bill King <bill.king@nokia.com>