Squish: Improve hook-into sub-processes
With this patch it's possible to not only hook into the
QmlApplicationViewer (although nothing else is currently done)
and additionally to not only send the closing event to the
sub-process. You can now define a function that will be executed
on the sub-process.
Both modified tests show an example how to use it.

Change-Id: I39b9959f2cf1d519b8afeb0c479ac2d68ea20ca6
Reviewed-by: Bill King <bill.king@nokia.com>
Reviewed-by: Robert Löhning <robert.loehning@nokia.com>