Avoid setting large minimum size for output panes.
The search result pane set a minimum size for the "New Search"
panel which forced this relative big minimum size on all panels
and made it impossible to make the output pane area small.
Now, the new search is wrapped into a scroll area, doesn't
enforce a minimum size anymore, and, instead, when initiating
a new search (via shortcut or menu action) resizes the pane
so the panel fits nicely.

Change-Id: I9afe67e66f324111a10a97f33c203846231e6a93
Reviewed-by: hjk <qthjk@ovi.com>