Cropping the examples screen shots to "Areas of interest".
The cropping coordinates are stored in a file
images_areasofinterest.xml. A tool to create/maintain the cropping
coordinates is included in this patch.

Then changes to Delegate.qml prevent QML from scaling the image.
If scaling is needed it is now the job of the Image provider.
dropshadow.png had to be adjusted because the border had different
widths and heights.

Change-Id: Ib25db896d40eeec0e0a373c971931dc5cabf3cbe
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@nokia.com>