Use QUuids to identify tool chains
Save/Restore has been updated to reflect this:

We now save manual as well as auto-detected tool chains into the user
toolChains.xml file.

When loading we load the SDK as well as the user toolChains.xml files.
We then do a autodetection run. All the tool chains restored that are
marked up as auto-detected are then checked again:
 * the tool chain was autodetected again: Throw away the newly auto-
   detected tool chain and reuse the saved one. This makes sure we
   keep the QUuid that is now part of the tool chains id.
 * the tool chain was not autodetected again: Throw away the saved
   tool chain.
 * the tool chain was newly autodetected: Add the new tool chain

We keep the old id around in the legacyId() method. It is used when
loading old .user files only. So there is no need to implement this
method for new tool chains.

Task-number: QTCREATORBUG-6702

Change-Id: Ifc0a216e5351a8a57db03615ba329e355e7d5f59
Reviewed-by: Daniel Teske <daniel.teske@nokia.com>