Add spaces between macros and string literals
Commit fa66a540 fixed string literal and macro concatenation
some time ago, but while it changed "literal1"M1"literal2" to
"literal1" M1 "literal2" it didn't touch M"literal". This patch
changes this to M "literal" for consistency.

This patch is the remnant of a cherry-pick of a C++11 compilation
fix equivalent to fa66a540 from KDAB's branch of QtC 2.4.0.

Change-Id: I87644e888a7e74b18d4b1ca10786cdc3fa798453
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>
Reviewed-by: hjk <qthjk@ovi.com>