VCSmanager: Fix QTC_ASSERTs
These tests are needed to make sure no broken data ends up in the cache,
causing inconsistencies in the cache. The asserts did a good job there,
catching a couple of issue, so I would like to keep them the way they are.

QDir is not needed at all as the data _should_ be perfectly fine already
anyway and should not be used if the asserts are not evaluated.

Change-Id: Icf113f8b02b6c7c7ba7498386b8465b2644daae8
Reviewed-by: Tobias Hunger <tobias.hunger@nokia.com>