Add a new option to main.cpp to specify a specific pid.
Add a new option to main.cpp which allows the user to specify a specific pid to
use to open a new file and forwards all the arguments to that pid.  Required some
changes to QtSingleApplication, but I did not alter any existing API.  Just added
some new methods for querying an instance of a specific pid.

Together with some changes to the terminal plugin this will allow the user to open
any file with Qt Creator from within the terminal plugin by simply invoking the
qtcreator binary.  All that is necessary is the addition of an alias like so:

if [ -n "$QTCREATOR_PID" ] ; then alias qtcreator="/path/to/qtcreator/binary/qtcreator -pid $QTCREATOR_PID" ; fi

This is entirely analagous to what kate allows where you can open any file with
kate from within the embedded terminal.

Change-Id: I476d78d673ee60052a02eb974eefc5368d24193b
Reviewed-by: Eike Ziller <eike.ziller@nokia.com>
Reviewed-by: hjk <qthjk@ovi.com>