Send the outputpane the visibilityChanged(true) state when icon is toggled.
When an outputpane is toggled visible/invisible/visible by clicking on the icon
itself the plugin is not made aware of the visiblityChanged(true) state since
changePage does not detect that the index has changed because indeed it has not.

This change just explicitly adds the visibilityChanged call after the toggle so
that the plugin is made aware.

Change-Id: I6e782f52c9ac926c39acd91b5533009899a4ec2a
Reviewed-by: hjk <qthjk@ovi.com>