Debugger: Fix JS debugging on Harmattan PR1.2
So far we assumend that the JSDebugAgent service switches to
the Enabled state after the constructor has run. That's apparently
not true for Harmattan PR1.2 (maybe the app / Qt calls process
events by hand, maybe it was a race condition from the start).

Fix this by checking the status already in the constructor, and
register ourselves as debugger agent if status == Enabled.

Change-Id: Icae189102116c8b572e05ce6c64f14b8f428ed2f
Reviewed-by: hjk <qthjk@ovi.com>
Reviewed-by: Alessandro Portale <alessandro.portale@nokia.com>