Fix a crash when loading the qt4projectmangerplugin.
This crash can happen when enabling/disabling a certain combination
of plugins and restarting.

There was a null pointer de-reference that was occurring because the
qt4projectmanagerplugin was referencing an action that it believed
to have been registered by the texteditorplugin when it was initialized.

However, apparently the texteditorplugin was not initializing its actions
at plugin initialization, but rather when 'extensionsInitialized' was called.
I do not know the call graph for when this is to be called, but I encountered
at least one situation where the qt4projectmanagerplugin was being initialized
before this.

Change-Id: Iede1831e0ac9c92b80a079157e1bdc8c66473470
Reviewed-by: Leandro Melo <leandro.melo@nokia.com>
Reviewed-by: Eike Ziller <eike.ziller@nokia.com>