React to qt versions changes on the project setup page.
Since the project setup now hapens in the projects mode, we need to
consider the user using the qt options dialog.

The TargetSetupPage owns the temporary qt versions created for
importing. The individual Qt4TargetSetupWidgets hold pointers
to those temporary qt version. Thus the TargetSetupPage needs
to inform the widgets when a pointer becomes invalid or a
already existing id needs to be replaced by a temporary qt
version. The interface of Qt4TargetSetupWidget is now
pretty bad and doesn't really work.

Task-Number: QTCREATORBUG-7020

Change-Id: I08799e482cb8f7f7b86c1bc14a0aabd444eee5d8
Reviewed-by: Tobias Hunger <tobias.hunger@nokia.com>