Fix compilation of QmlJS code formatter test.
Done by moving ITextMark and ITextMarkable from itexteditor.h
to itextmark.h. The test uses parts of the texteditor plugin by
compiling some of its files. For that to work, the code has to be
fairly stand-alone. The reference to EditorManager in
itexteditor.cpp broke it.

Long term, the test should become a plugin based test to avoid this
kind of problem.

Change-Id: Ibae3bcd7365eb4e2814cd07011cad0180507e82b
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@nokia.com>