Fix non-stop mode in gdb when you click continue
When you hit a brekpoint in gdb non-stop mode, if you click continue in the
UI the debugger was killed if the current thread is not the one being
stopped. This is anoying as gdb does not change current thread
automatically. The fix shows the error but does not kill the inferior.
Fixed style issues again.

Change-Id: I1a2e484108e936581c7b82c985f2b057263e3d6b
Reviewed-by: hjk <qthjk@ovi.com>