Fix non-stop mode in gdb when you click continue
When you hit a brekpoint in gdb non-stop mode, if you click continue in the
UI the debugger was killed if the current thread is not the one being
stopped. This is anoying as gdb does not change current thread
automatically. The fix shows the error but does not kill the inferior.
Fixed style issues again.

Reviewed-by: hjk <qthjk@ovi.com>
(cherry picked from commit 47def926c7172dfff570cfc9b7192e201c0ee3d3)

Change-Id: I18e8b77b615cfceec402b140337709074ad96dc5
Reviewed-by: hjk <qthjk@ovi.com>