fix non-stop mode for executeNext and executeStep
Following the  Change-Id: I18e8b77b615cfceec402b140337709074ad96dc5, I have
found the same behaviour in handleExecuteNext and HandleExecuteStep than in
hanleExecuteContinue. The fix is the same, If you click one of the two
buttons while having a different thread selected, notify the GUI of the gdb
error message but do not kill the debugger neither the inferior as the situation
can be addressed by clicking again step or next after switching to the
stopped thread.

Change-Id: I9f2951ea9846621156ceb3998514d7667d59e7af
Reviewed-by: hjk <qthjk@ovi.com>