Don't resolve symlinks in browser for qmake executable.
On Mac, the Qt packages install qmake to "/usr/bin/qmake-4.7" etc, and
put a symlink "qmake" in a browsable place. If the browser resolves
symlinks, we would need to accept all these different qmake naming variants
instead of just "qmake". Also, there is no reason for resolving the symlink,
in the worst case (like in this case) it just makes life harder.

Task-number: QTCREATORBUG-6222
Change-Id: I973a648ea2d1586075723efa6164f44d8fd3ecad
Reviewed-by: Daniel Teske <daniel.teske@nokia.com>