QmlDesigner: crash fix
It was a very stupid idea to call QApplication::processEvents(); here
to avoid hickups.

The processEvents() forces the whole designDocumentController managment to be
be reentrant which it clearly is not (and should not be).
The reason is simply that resetView is called from there.
Adding a "flag" for the processEvents is also not a reasonable option.
This was just very bad style.
So I remove it.

Task-number: QTCREATORBUG-7120
Change-Id: I111cf9421f63b9ce44488f0d274624ff48777cef
Reviewed-by: Kai Koehne <kai.koehne@nokia.com>