Curly-braces in "Help -> Search for..." crashes Creator.
Task-number: QTCREATORBUG-6212

We need to escape special characters that are used inside the
search engine as field delimiter. As soon as the search engine
proccesses the search string it will remove the character and
we might end up with an invalid string, forcing it to throw an
exception thats not catched.

Change-Id: I8b66c87c3137a1f175ead5df85c7f53fdcb5193e
Reviewed-by: Eike Ziller <eike.ziller@nokia.com>