Device: Add create method to IDeviceFactory
Add a create method to the IDeviceFactory and make it replace the
createWizard method. The create method may or may not pop up a wizard.

Remove IDeviceWizard as suggested by Christian.

This patch also adds a canCreate() method to check whether a factory should
be asked to create something or not.

Change-Id: Iaf16aa407530022e8f3804093b58dc3f120f7599
Reviewed-by: Christian Kandeler <christian.kandeler@nokia.com>