RemoteLinux: Fix ports gatherer API.
The ports gatherer used to take an SshConnection for historical reasons
(connection sharing in the absence of a connection manager). This is no
longer required, since all the information needed for creating or
re-using a connection is available from the device.
In addition, the old code assumes the connection is already established,
which some newer callers did not adhere to, so this patch also fixes a
bug.

Change-Id: I3fd7fec7de9b64126358749f727a403bfa1e0a94
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@nokia.com>